backend/orders/requirements.txt
flask
flask-jwt-extended
flask-cors
psycopg2-binary
pika
prometheus-flask-exporter